+++ app/javascript/vue/constants/maps.js
@@ -1,1 +1,2 @@
 export const MAP_SHAPE_AGGREGATE = 'AgreggateShape'
+export const MAP_SHAPE_ASSERTED_DISTRIBUTION_ABSENT = 'AssertedAbsentShape'